python/ai-wallet-reputation-nft/requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
web3>=6.0.0
requests>=2.25.0 
openai>=1.0.0 